Print warnings for unbooted slices in console (#349)
Add another console “plugin” to overload `Slice.keys` and print a one-time-per-slice warning when accessing keys for an unbooted slice.

This “plugin” is slightly differently-shaped to the `SliceReaders` one we have currently, but it needs to be because its purpose is not actually to enhance the console context, but rather to patch some external code. I think this difference is fine, and in the future we can move to make a more formal definition of what it means to be a console plugin.